// client/src/interfaces/converted.ts
export default interface converted {
    fileName: string;
    id: number;
    inputId: string;
    resolution: string;
    isThumbnail: number;
    key: string;
    location: string;
}
export interface convertedImageType {
    id: number;
    inputId: string;
    fileName: string;
    key: string;
    location: string;
    page: number;
}